Reorganizing of the WPFUtilities Project.
