+++ CyberCavsScoutingSystem/CCSSDomain/GameEditingData.cs
@@ -3,6 +3,7 @@ using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Windows.Media;
 using WPFUtilities.Validation;
+using WPFUtilities.Validation.Inputs;
 
 namespace CCSSDomain;
