+++ CyberCavsScoutingSystem/GameMaker/App.xaml
@@ -7,6 +7,7 @@
 			 xmlns:converters="clr-namespace:GameMaker.Converters"
 			 xmlns:wpfUtilities="clr-namespace:WPFUtilities;assembly=WPFUtilities"
 			 xmlns:domain="clr-namespace:CCSSDomain;assembly=CCSSDomain"
+			 xmlns:extensions="clr-namespace:WPFUtilities.Extensions;assembly=WPFUtilities"
 			 Startup="ApplicationStartup">
 	<Application.Resources>
 
@@ -70,7 +71,7 @@
 						<TextBlock Text="Description stuff here."/>
 						<ItemsControl ItemsSource="{Binding ValidationErrors}">
 							<ItemsControl.ItemTemplate>
-								<DataTemplate d:DataType="{wpfUtilities:GenericObjectFactory Type={x:Type wpfUtilities:ValidationError`1}, T=domain:ErrorSeverity}">
+								<DataTemplate d:DataType="{extensions:GenericObjectFactory Type={x:Type wpfUtilities:ValidationError`1}, T=domain:ErrorSeverity}">
                                     <DataTemplate.Resources>
 										<Style TargetType="TextBlock">
 											<Setter Property="Foreground"
@@ -90,7 +91,7 @@
 
 		<Style x:Key="MultiStringInputBorderStyle" TargetType="Border">
 			<d:Style.DataContext>
-                <d:DesignInstance Type="{wpfUtilities:GenericObjectFactoryExtension2 Type={x:Type wpfUtilities:ValidationError`1}, T1=domain:VersionNumber, T2=domain:ErrorSeverity}"/>
+                <d:DesignInstance Type="{extensions:GenericObjectFactoryExtension2 Type={x:Type wpfUtilities:ValidationError`1}, T1=domain:VersionNumber, T2=domain:ErrorSeverity}"/>
                 <!--<d:DesignInstance Type="wpfUtilities:MultiStringInput, IsDesignTimeCreatable=True"/>-->
 			</d:Style.DataContext>
 
@@ -104,7 +105,7 @@
 						<TextBlock Text="Description stuff here."/>
 						<ItemsControl ItemsSource="{Binding CovalidationErrors}">
 							<ItemsControl.ItemTemplate>
-								<DataTemplate d:DataType="{wpfUtilities:GenericObjectFactory Type={x:Type wpfUtilities:ValidationError`1}, T=domain:ErrorSeverity}">
+								<DataTemplate d:DataType="{extensions:GenericObjectFactory Type={x:Type wpfUtilities:ValidationError`1}, T=domain:ErrorSeverity}">
                                     <DataTemplate.Resources>
 										<Style TargetType="TextBlock">
 											<Setter Property="Foreground"
