+++ CyberCavsScoutingSystem/GameMaker/Converters/ErrorToBrushConverters.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Windows.Media;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using CCSSDomain;
-using WPFUtilities;
+using WPFUtilities.Validation.Errors;
 
 namespace GameMaker.Converters;
