+++ CyberCavsScoutingSystem/WPFUtilities/ReadOnlyList.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-
 using WPFUtilities.Extensions;
 
 namespace WPFUtilities;
