+++ CyberCavsScoutingSystem/WPFUtilities/Validation/Delegates/Validators.cs
@@ -1,0 +1,15 @@
+﻿using WPFUtilities.Validation.Errors;
+
+namespace WPFUtilities.Validation.Delegates;
+
+
+
+public delegate ReadOnlyList<ValidationError<TSeverityEnum>>
+	InputValidator<in TOutput, TSeverityEnum>
+	(TOutput targetObject)
+	where TSeverityEnum : ValidationErrorSeverityEnum<TSeverityEnum>, IValidationErrorSeverityEnum<TSeverityEnum>;
+
+public delegate ReadOnlyList<ValidationError<TSeverityEnum>>
+	InputValidator<in TOutput, in TValidationParameter, TSeverityEnum>
+	(TOutput targetObject, TValidationParameter parameter)
+	where TSeverityEnum : ValidationErrorSeverityEnum<TSeverityEnum>, IValidationErrorSeverityEnum<TSeverityEnum>;
