+++ CyberCavsScoutingSystem/WPFUtilities/Validation/ValidationEvent.cs
@@ -1,5 +1,5 @@
-﻿using System.Collections.Generic;
-using System;
+﻿using System;
+using System.Collections.Generic;
 
 namespace WPFUtilities.Validation;
