+++ CyberCavsScoutingSystem/WPFUtilities/Validation/ValidationTrigger.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Linq;
+using WPFUtilities.Validation.Delegates;
+using WPFUtilities.Validation.Errors;
 
-namespace WPFUtilities.Validation; 
+namespace WPFUtilities.Validation;
